- PostImageId Int -> Long 으로 타입변경 -

+++ src/main/java/org/nurim/nurim/domain/entity/PostImage.java
@@ -15,7 +15,7 @@ public class PostImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "postImageId", nullable = false)
-    private int postImageId;
+    private Long postImageId;
 
     @ManyToOne()
     @JoinColumn(name = "postId")
